Add files via upload
This file is not yet fully working. The problem is with the NetworkX adjacency_matrix() function returning a sparse SciPy array, while the documentation says it should return a 2d NumPy array.

This implements the algorithm for splitting a network based on the specified paper. It will soon be updated to include splitting networks into more than 2 communities.